fix(watch): Search both stdout and stderr for output
Sometimes, depending on the scenario, I've found the output text
that I want in both stdout and stderr, and now always at the time
I was expecting.

So with this change, I check the one I'm expecting first, and if nothing
is there, then we grab the other one just to be safe.  Because one of these
is almost always empty (or so I have seen).