fix(watch): Don't break on sub command failure
Originally, I thought the code should break when any sub command failed,
hypothesizing that the most common case is where the later stringed commands depend on
the earlier ones.  But I think that is too presumptive here.

The user will be able to see if each command fails, and then given their specific
context, should be able to understand that OF COURSE the second one failed,
because the first one already failed.