fix(watch): Don't err out if second file changes
Previously, it did a very poor job of handling the build process
if a second file changed before the original build was finished.

This should be resolved now, with the builds all happening in the
order they were detected.

This logic has a lot of room for improvement though.  It could cancel
the previous build perhaps if another change were made to that same
package, but canceling it would be a lot of work because each stage
in the process would have to know how to cancel itself.

---

I'm also thinking that it would be really cool to display all the packages that were going to be built (like the summary at the end) before they
are built.  All the packages would have spinners, and then when they
finished, the spinner would turn into the check.