fix(watch): Only have chokidar ignore node_modules
It was previously also told to ignore certain typical build
directory names and dot files.  This ignoring was done to
avoid unnecessary executions if only the built files changed.

However, I think this ignoring is the root cause behind issue
subdirectory of the packages/ dir.  Any change to any sub directory
of "libs" was not triggering an event from chokidar, and
additionally, any package with "lib" in the name was also not
causing a change event.

Because unorepo will, I think, in most cases be the process
that is updating the built files, this should not really be an
issue.

---

A better approach might be to allow that user to specify which
files should be ignore, that could be a future enhancement.